spectra, spectrum and hcube added to data
