+++ libpysat/transform/baseline_code/ccam_remove_continuum.py
