+++ libpysat/transform/baseline_code/common.py
