+++ libpysat/transform/baseline_code/dietrich.py
