+++ libpysat/transform/baseline_code/kajfosz_kwiatek.py
