+++ libpysat/transform/baseline_code/mario.py
