+++ libpysat/transform/baseline_code/rubberband.py
