+++ libpysat/transform/baseline_code/spl_init.py
