+++ libpysat/transform/baseline_code/watrous.py
