+++ libpysat/transform/baseline_code/wavelet.py
